Avoid using the relative path to the server executable in client guide
It's not necessary that the server executable is located and named
always the same. We may try to hijack this using various CMake
OUTPUT tricks, but it's easier to utilize the actual file path in
the client executable using definition.

Amends dedd3edd3feb49c285985797dbd459ba35e64b8c

Pick-to: 6.8 6.9
Task-number: QTBUG-125406
Change-Id: Ice60e22caa6c5c14024f57fe133f2268ad96f574
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>