Doc: Split up protobuf attributions between tools and tests
We should aim to have the attributions as close as possible to the
actual sources. The old 'combined' attribution also did mix up the
in-source copy of some protobuf files, compared to the dependency
that is provided by provisioning.

This also fixes an issue in our documentation infrastructure, where
we remove tests/ directory before building the Qt Framework
attributions. This generated a warning so far:

     File ./qtgrpc/qt_attribution.json: Directory '/home/vagrant/qt-everywhere-src-6.8.1/qtgrpc/tests/auto/protobuf/conformance/data' does not exist.

Pick-to: 6.8 6.9
Change-Id: I26c8483737ab6a6e863b3ee4a683c77cb9af5c0a
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>